fix(form-analysis-service): resolver ambigüedad Log.debugf con cast explícito a long
Long es ambiguo entre overloads debugf(String,long,long) y
debugf(String,Object,Object). Cast explícito (long) fuerza el overload correcto.

+++ form-analysis-service/src/main/java/es/kitti/formanalysis/service/FormAnalysisService.java
@@ -59,7 +59,7 @@ public class FormAnalysisService {
         List<FlagResult> rulesFlags = formAnalysisRules.evaluate(event);
         String prompt = llmPromptBuilder.build(event);
 
-        Log.debugf("Calling LLM for request %d", event.adoptionRequestId());
+        Log.debugf("Calling LLM for request %d", (long) event.adoptionRequestId());
         return llmClient.analyzeTextFields(prompt)
                 .onFailure().recoverWithItem(e -> {
                     Log.warnf("LLM failed for request %d (%s): %s",
@@ -68,7 +68,7 @@ public class FormAnalysisService {
                 })
                 .onItem().transform(raw -> {
                     Log.debugf("LLM responded for request %d: %d chars",
-                            event.adoptionRequestId(), raw != null ? raw.length() : 0);
+                            (long) event.adoptionRequestId(), (long) (raw != null ? raw.length() : 0));
                     return raw;
                 })
                 .onItem().transform(this::parseLlmResponse)
@@ -106,7 +106,7 @@ public class FormAnalysisService {
                     }).toList();
 
                     Log.debugf("Persisting analysis for request %d: decision=%s critical=%d warning=%d",
-                            event.adoptionRequestId(), decision, criticalCount, warningCount);
+                            (long) event.adoptionRequestId(), decision, criticalCount, warningCount);
                     return persistenceService.persist(analysis, formFlags)
                             .onItem().transformToUni(saved -> {
                                 Log.infof("Form analysis completed for request %d: %s",
